+++ src/test/extension.test.ts
@@ -1,0 +1,10 @@
+// import * as assert from 'assert';
+// import * as vscode from 'vscode';
+// import * as extension from '../extension';
+
+suite('Extension Test Suite', () => {
+  // test('Sample test', () => {
+  //   assert.strictEqual(-1, [1, 2, 3].indexOf(5));
+  //   assert.strictEqual(-1, [1, 2, 3].indexOf(0));
+  // });
+});